+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## Unreleased
 
+### Added
+
+* All the public API names are now correctly marked `public` in Julia 1.11 and above. ([#83])
+
 ### Changed
 
 * The string `repr` of `DatasetVersion` (e.g. `dataset.versions`) is now valid Julia code. ([#84])
@@ -145,4 +149,5 @@ Initial package release.
 [#53]: https://github.com/JuliaComputing/JuliaHub.jl/issues/53
 [#58]: https://github.com/JuliaComputing/JuliaHub.jl/issues/58
 [#74]: https://github.com/JuliaComputing/JuliaHub.jl/issues/74
+[#83]: https://github.com/JuliaComputing/JuliaHub.jl/issues/83
 [#84]: https://github.com/JuliaComputing/JuliaHub.jl/issues/84