+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## Version [v0.1.12] - 2025-04-25
 
 ### Added
 
@@ -134,6 +134,7 @@ Initial package release.
 [v0.1.9]: https://github.com/JuliaComputing/JuliaHub.jl/releases/tag/v0.1.9
 [v0.1.10]: https://github.com/JuliaComputing/JuliaHub.jl/releases/tag/v0.1.10
 [v0.1.11]: https://github.com/JuliaComputing/JuliaHub.jl/releases/tag/v0.1.11
+[v0.1.12]: https://github.com/JuliaComputing/JuliaHub.jl/releases/tag/v0.1.12
 [#1]: https://github.com/JuliaComputing/JuliaHub.jl/issues/1
 [#2]: https://github.com/JuliaComputing/JuliaHub.jl/issues/2
 [#3]: https://github.com/JuliaComputing/JuliaHub.jl/issues/3